+++ libs/attrs-utils/src/thds/attrs_utils/cattrs/__init__.py
@@ -4,7 +4,9 @@ __all__ = [
     "DEFAULT_STRUCTURE_HOOKS",
     "DEFAULT_UNSTRUCTURE_HOOKS_JSON",
     "PREJSON_UNSTRUCTURE_COLLECTION_OVERRIDES",
+    "DisallowedConversionError",
     "default_converter",
+    "format_cattrs_classval_error",
     "setup_converter",
 ]
 
@@ -16,6 +18,7 @@ from .converter import (
     default_converter,
     setup_converter,
 )
+from .errors import DisallowedConversionError, format_cattrs_classval_error
 
 DEFAULT_JSON_CONVERTER = setup_converter(
     default_converter(),